+++ pro/ai/mcp/client/src/main/kotlin/org/http4k/ai/llm/tools/McpLLMTools.kt
@@ -1,9 +1,12 @@
 package org.http4k.ai.llm.tools
 
+import dev.forkhandles.result4k.Result
 import dev.forkhandles.result4k.flatMap
 import dev.forkhandles.result4k.map
 import dev.forkhandles.result4k.mapFailure
 import org.http4k.ai.llm.LLMError
+import org.http4k.ai.llm.LLMResult
+import org.http4k.ai.mcp.McpResult
 import org.http4k.ai.mcp.ToolResponse.Ok
 import org.http4k.ai.mcp.client.McpClient
 import org.http4k.ai.mcp.model.Meta
@@ -20,7 +23,7 @@ class McpLLMTools(private val client: McpClient) : LLMTools {
         .map { it.map(McpTool::toLLM) }
         .mapFailure { LLMError.Internal(Exception(it.toString())) }
 
-    override fun invoke(request: ToolRequest) =
+    override fun invoke(request: ToolRequest): LLMResult<ToolResponse> =
         client.tools()
             .call(request.name, McpToolRequest(request.arguments, meta = Meta(request.id.value)))
             .mapFailure { it.toLLM() }