+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 This list is not intended to be all-encompassing - it will document major and breaking API
 changes with their rationale when appropriate:
 
+### v6.29.0.0 (uncut)
+
+- **http4k-***: Upgrade versions
+- **http4k-ai-mcp-client**: [Breaking] Alteration to ToolResponse.Error to have content instead of just a string
+- **http4k-ai-mcp-sdk**: [Breaking] Alteration to ToolResponse.Error to have content instead of just a string
+
 ### v6.28.1.0
 - **http4k-***: Upgrade versions
 - **http4k-ai-mcp-testing**: Remove unnecessary dependencies, various fixes to the MCP App test host